migrate bag extraction to its own file
